requirements.txt
alembic==1.7.5
aniso8601==9.0.1
anyio==3.4.0
apispec==5.1.1
apispec-webframeworks==0.5.2
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
flake8==4.0.1
flasgger==0.9.5
Flask==2.0.2
Flask-Alembic==2.0.1
Flask-JWT-Extended==4.3.1
Flask-Limiter==2.1
Flask-Login==0.5.0
Flask-OpenTracing==1.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
gevent==21.12.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
jaeger-client==4.8.0
Jinja2==3.0.3
jsonschema==4.4.0
limits==2.3.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
mistune==2.0.2
mypy-extensions==0.4.3
opentracing==2.4.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.19.3
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pycryptodomex==3.12.0
pydantic==1.8.2
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.2.5
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
threadloop==1.0.2
thrift==0.15.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
Werkzeug==2.0.2
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
